Move settings files to a settings folder
