+++ example/lib/ui/random_value_page.dart
@@ -2,10 +2,6 @@ import 'dart:math';
 import 'package:example/utils.dart';
 
 import 'package:flutter/material.dart';
-import 'package:sleek_circular_slider/configuration/CircularSliderColors.dart';
-import 'package:sleek_circular_slider/configuration/CircularSliderGeometry.dart';
-import 'package:sleek_circular_slider/configuration/CircularSliderShadow.dart';
-import 'package:sleek_circular_slider/configuration/CircularSliderText.dart';
 import 'package:sleek_circular_slider/sleek_circular_slider.dart';
 
 class RandomValuePage extends StatefulWidget {
