+++ lib/sleek_circular_slider.dart
@@ -1,4 +1,9 @@
 library sleek_circular_slider;
 
 export 'src/circular_slider.dart';
-export 'src/settings.dart';
+export 'src/settings/CircularSliderSettings.dart';
+export 'src/settings/CircularSliderFeatures.dart';
+export 'src/settings/CircularSliderColors.dart';
+export 'src/settings/CircularSliderGeometry.dart';
+export 'src/settings/CircularSliderShadow.dart';
+export 'src/settings/CircularSliderText.dart';
