+++ lib/src/settings/CircularSliderColors.dart
