+++ lib/src/settings/CircularSliderFeatures.dart
