+++ lib/src/settings/CircularSliderGeometry.dart
