+++ lib/src/settings/CircularSliderShadow.dart
