+++ lib/src/settings/CircularSliderText.dart
